Use a central fixture for the ES client
