--- tests/common/fixtures/__init__.py
@@ -1,3 +1,0 @@
-from tests.common.fixtures.elasticsearch import es_client, init_elasticsearch
-
-__all__ = ("es_client", "init_elasticsearch")
